Version 1.3.0, Added: - User profile section - Wishlists section Fixed: - Seach section and navigation

+++ lib/data/services/product_service.dart
@@ -117,7 +117,7 @@ class ProductService implements IProductService {
 
       final productData = productDoc.data()!;
       final String category = productData['category'];
-      final double price = productData['price'];
+      final int price = productData['price'];
 
       final QuerySnapshot similarProducts = await _firestore
           .collection('products')
